Understanding Prop Chain Problems
It pass down props across multiple levels, for example, the persons component only forward the authenticated status as a props but it doesn't really care. This simple leads to extra redundancy and makes components a bit less reusable Because wherever we're using the persons component, we have to make sure we pass authenticated status so that we can forward it.